requirements.txt
asgiref==3.8.1
Django==5.2.1
et_xmlfile==2.0.0
lxml==5.4.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
